create LinkCopyComponent to display and copy poll voting link
I created a LinkCopyComponent. This component accepts a link prop.
It displays this link prop and truncates it with ellipses when the link is longer than the space available.
It has a copy button on the right that copies the link in the link prop to the clipboard when clicked.
When the copy button is clicked, the text changes to copied.
Using setTimeout(), I reset the text on the button back to copy.